fix the error fields and adapt the new fields in new schema prisma to nice fields in this service

+++ src/village/village.service.ts
@@ -11,7 +11,7 @@ export class VillageService {
 
         if (user_id == null) throw new UnauthorizedException("Failed to get the user id from token!")
 
-        const find_unique_name = await this.prisma.village.findFirst({
+        const find_unique_name = await this.prisma.village.findUnique({
             where: {
                 Name: data.Name
             }
@@ -31,7 +31,7 @@ export class VillageService {
                 }
             })
 
-            if (!data_create) throw new BadRequestException("Failed to create because the data is nill!")
+            if (!data_create || data_create == null && data_create == undefined) throw new BadRequestException("Failed to create because the data is nill!")
 
             return data_create
 
@@ -45,6 +45,8 @@ export class VillageService {
 
         if (user_id == null) throw new UnauthorizedException("Failed to get the user id from token!")
 
+        if (id == null && id == undefined) throw new BadRequestException("Failed to detect the id from the parameter!")
+
         try {
 
             const delete_data = await this.prisma.village.delete({
@@ -53,7 +55,7 @@ export class VillageService {
                 }
             })
 
-            if (!delete_data) throw new BadRequestException("Not Found!")
+            if (!delete_data || delete_data == null && delete_data == undefined) throw new BadRequestException("Can't find the data that you want to delete it!")
 
             return true
 
@@ -66,6 +68,8 @@ export class VillageService {
     async updateVillage(data: any, user_id: number, id: number) {
 
         if (user_id == null) throw new UnauthorizedException("Failed to get the user id data from token!")
+
+        if (id == null && id == undefined) throw new BadRequestException("Failed to detect the id in the parameter request!")
 
         //tools
         const update_data = CheckIsNullWithNumber(data)
@@ -98,9 +102,21 @@ export class VillageService {
 
         try {
 
-            const getAll_village = await this.prisma.village.findMany()
+            const getAll_village = await this.prisma.village.findMany({
+                select: {
+                    Name: true,
+                    Address: true,
+                    Village_Age: true,
+                    Leader_Village: {
+                        select: {
+                            Username: true,
+                            Address: true
+                        }
+                    }
+                }
+            })
 
-            if (!getAll_village) throw new BadRequestException("Failed to get village!")
+            if (!getAll_village || getAll_village == null && getAll_village == undefined) throw new BadRequestException("Failed to get village!")
 
             return getAll_village
